wip, mewgo
mewgo is a rewrite of the ball of python project display and management
portion of MeowCAD in Go(lang).  Using the Iris framework.  It's unclear
if it's really going to be significantly faster (and be able to resist
the hug of love) but it's more mangeable and extensible at least.